Adjusted Equation values (deadline)

+++ StishBoard/StishBoard/Computer.cs
@@ -23,9 +23,9 @@ namespace StishBoard
             StishMiniMaxNode GameNode = new StishMiniMaxNode(null, StishBoard.Instance.Player1);
             GameNode.NodeBoardState = new BoardState(StishBoard.Instance);
 
-            //int evaluation = MiniMaxMind.Instance.BuildABTree(GameNode, 4, int.MinValue, int.MaxValue, -1);
+            //double evaluation = MiniMaxMind.Instance.BuildABTree(GameNode, 4, int.MinValue, int.MaxValue, -1);
             MiniMaxMind.Instance.RecBuildMMTree(GameNode, 4);
-            int evaluation = MiniMaxMind.Instance.TraverseTree(GameNode, 4, -1);
+            double evaluation = MiniMaxMind.Instance.TraverseTree(GameNode, 4, -1);
 
             //ForeSight.Instance.PredctionCount();
